requirements.txt
Django
djangorestframework
psycopg2-binary
dj-database-url
django-environ==0.9.0
whitenoise